Image Capture: Extract to separate class (no functional change)
Aids readability.
